fix: use 64-bit float functions for compatible behavior
Even though opus works with floats (32-bit), it still uses a bunch of standard math functions operating on doubles (64-bit). This is because C89 doesn't require the float variants to be present (there is `exp`, but no `expf`). This was later remedied in C99.

In previous efforts for delibcfication I've replaced those double functions with rust native calls to methods on f32, which compile down to those float functions (like `f32::exp` -> `expf`). This prevented me from getting the exact same encoder behavior, because, apparently, not all of the functions behave the same when wrapped with a conversion to double and back. For some experimentation - see https://gist.github.com/DCNick3/f929bdf96e71a817ab0954c05172da35.

This commit brings back the conversions to f64 when using most of the float points functions, except for abs and floor which seem to behave the same way. `sqrt` also seems to behave the same way in my experiments, but I've changed it just to be safe.

Unlike the way it was before, we now have math functions in celt's mathops module which correspond to macros used throughout its codebase. This way the f32->f64->f32 conversions become much less copy-pasty